requirements.txt
streamlit>=1.36
pandas>=2.2
numpy>=1.26
scikit-learn>=1.4
spacy>=3.7
rapidfuzz>=3.9
pypdf>=4.2
python-docx>=1.1
altair>=5.2
rich>=13.7
